+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+- BREAKING: Make build context proxy create the context internally
+
 ### Added
 
 - Make proxy object possible to extend an existing object